Migrate images and attachments
The first image attachment will be used as the image of the part. All
other attachments will be uploaded as attachments to that part.

The InvenTree python module requires a local file to obtain the
filename, so each file is only downloaded when needed into a temporary
directory and removed afterwards.